US3T3: move Friend, FriendRequestError, RelationshipStatus into Models
