+++ PodConnect/Repositories/FriendRepository.swift
@@ -122,33 +122,3 @@ class FriendRepository {
         }
     }
 }
-
-struct Friend: Codable {
-    @DocumentID var id: String?
-    let user1UID: String
-    let user2UID: String
-    let since: Date
-}
-
-enum FriendRequestError: LocalizedError {
-    case cannotAddSelf
-    case alreadyFriends
-    case requestAlreadySent
-
-    var errorDescription: String? {
-        switch self {
-        case .cannotAddSelf:
-            return "You can't send a friend request to yourself."
-        case .alreadyFriends:
-            return "You are already friends with this user."
-        case .requestAlreadySent:
-            return "A friend request has already been sent to this user."
-        }
-    }
-}
-
-enum RelationshipStatus {
-    case none
-    case requestSent
-    case friends
-}
